Restructuring files. Using more ES6 syntax. Working towards more reactivity.
