+++ contentcuration/contentcuration/static/js/edit_channel/preview/hbtemplates/preview_manager.handlebars
