+++ contentcuration/contentcuration/static/js/edit_channel/preview/hbtemplates/preview_templates/tabs.handlebars
@@ -1,3 +1,5 @@
 {{#each presets}}
-    <li class="btn preview_btn_tab preview_file" value="{{id}}" title="{{#translate id}}{{/translate}}">{{#translate id}}{{/translate}}</li>
+    <li class="btn preview_btn_tab preview_file" value="{{id}}" title="{{#translate id}}{{/translate}}">
+      {{#translate id}}{{/translate}}
+    </li>
 {{/each}}
