+++ contentcuration/contentcuration/static/js/edit_channel/preview/models/preview_manager.js
@@ -1,0 +1,12 @@
+import { Model } from 'backbone';
+
+export default Model.extend({
+    // probably not needed, but mark down what we got.
+    // TODO validate?
+    defaults: {
+      content_model: null,
+      current_preview: null,
+      previewView: null,
+      format_presets: [],
+    },
+  });
