--- contentcuration/contentcuration/static/js/edit_channel/preview/views.js
@@ -1,3 +1,0 @@
-export { default as PreviewModalView } from './child_views/preview_modal_view';
-export { default as PreviewView } from './child_views/preview_view';
-export { default as PreviewItemView } from './child_views/preview_item_view';
