+++ contentcuration/contentcuration/static/js/edit_channel/preview/views/index.js
@@ -1,0 +1,4 @@
+// maintaining old API
+export { default as PreviewModalView } from './child_views/preview_modal';
+export { default as PreviewView } from './child_views/preview_manager';
+export { default as PreviewItemView } from './child_views/preview';
